+++ node/rootwalker_test.go
@@ -95,9 +95,13 @@ func fillContainer1(c *Container, pk cipher.PubKey,
 		Thread{"Expressions", persons[2], posts2},
 		Thread{"Testing", persons[3], posts3},
 	)
-	r.InjectMany("Board",
-		Board{"Test", persons[3], dynPost, threads[2:]},
-		Board{"Talk", persons[1], dynPerson, threads[:2]},
+	r.Append(
+		r.MustDynamic("Board",
+			Board{"Test", persons[3], dynPost, threads[2:]},
+		),
+		r.MustDynamic("Board",
+			Board{"Talk", persons[1], dynPerson, threads[:2]},
+		),
 	)
 
 	return